+++ vitest.config.ts
@@ -1,4 +1,3 @@
-import swc from 'unplugin-swc';
 import tsconfigPaths from 'vite-tsconfig-paths';
 import { defineConfig } from 'vitest/config';
 
@@ -22,10 +21,5 @@ export default defineConfig({
     },
     setupFiles: ['./tests/setup.ts'],
   },
-  plugins: [
-    tsconfigPaths(),
-    swc.vite({
-      module: { type: 'es6' },
-    }),
-  ],
+  plugins: [tsconfigPaths()],
 });